[MultiQuickFix] minor API cleanup
